[refactor] 차트 애니메이션 확실하게 구현 완료, d3코드 분리 완료
