+++ src/page/mainPage.tsx
@@ -4,7 +4,7 @@ import SideMenuBar from '../components/mainPageComponent/sideMenu'
 import Group from '../components/mainPageComponent/groups'
 // import Search from '../components/mainPageComponent/search'
 import Typing from '../components/mainPageComponent/typing'
-import D3Canvas from '../components/mainPageComponent/d3'
+import D3Canvas from '../components/mainPageComponent/d3/d3'
 import LocationButton from '../components/mainPageComponent/locationButton'
 
 type Log = {
